// src/generated/models/New_internalordersModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const New_internalordersnew_orderstatus = {
  100000000: 'Submitted',
  100000001: 'Approved',
  100000002: 'InProgress',
  100000003: 'Ordered',
  100000004: 'Delivered',
  100000005: 'Denied'
} as const;
export type New_internalordersnew_orderstatus = keyof typeof New_internalordersnew_orderstatus;
export const New_internalordersstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type New_internalordersstatecode = keyof typeof New_internalordersstatecode;
export const New_internalordersstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type New_internalordersstatuscode = keyof typeof New_internalordersstatuscode;

export interface New_internalordersBase {
  importsequencenumber?: number;
  new_deliverylocation?: string;
  new_internalorderid: string;
  "new_Item@odata.bind": string;
  new_neededby?: string;
  new_notes?: string;
  new_orderdate?: string;
  new_orderid?: number;
  new_orderstatus?: New_internalordersnew_orderstatus;
  new_quantity?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: New_internalordersstatecode;
  statuscode?: New_internalordersstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface New_internalorders extends New_internalordersBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  new_itemname?: string;
  new_orderstatusname?: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  new_item?: object;
  _new_item_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}